Assignment-2/requirements.txt
bip-utils
python-bitcoinlib
requests